Fix gcc warnings and improve Dict reallocation breakpoints: now shrinks at a load factor of 20%, and expands at a load factor of 70%; in both cases the final load factor is set to LOAD_AFTER_REALLOC which is 0.40 atm

+++ libobject/src/strings_methods.c
@@ -140,7 +140,7 @@ t_bool  _string_match(const String *self, const char *compare)
 
 ssize_t  _string_nmatch(const String *self, const char *compare)
 {
-    return (self->contained && compare ? nmatch((char *)self->contained, compare) : -1);
+    return (self->contained && compare ? (ssize_t)nmatch((char *)self->contained, compare) : -1);
 }
 
 Object          *_string_split(const String *self, const Class *type, const char *sep)
